feat: Refactor core infrastructure. I have no idea how I thought the first design was "good". Attempting to update today was painful, and I only wrote the first one a few months ago...
