+++ package.json
@@ -22,6 +22,7 @@
     "date-fns": "^2.16.1",
     "got": "^11.8.1",
     "js-yaml": "^3.14.0",
+    "typedi": "^0.10.0",
     "typescript": "^4.0.5"
   },
   "devDependencies": {
