+++ src/Modules/RapidRecovery/ProtectedMachineSchema.ts
